extracted ammount selector sub widgets into seperate files
